requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
pinecone_client==2.2.1
pinecone_text==0.4.2
pydantic==1.10.8
python-dotenv==0.21.1
redis==4.6.0
sse_starlette==1.6.1
starlette==0.27.0
